+++ packages/brick_sqlite/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+* Revise `exists` in SqliteProvider to query with a SQL statement instead of hydrating whole models and associations
+* Add `selectStatement` to `QuerySqlTransformer`'s default constructor. When `true,` `statement` will begin `SELECT FROM`; when `false`, `statement` will begin `SELECT COUNT(*)`. Defaults `true`.
+
 ## 0.1.2
 
 * When searching with `doesNotContain` apply same fuzzy search `%` that `Compare.contains` enjoys